requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
python-dotenv
google-generativeai
pytest
